cleanUp

+++ app/src/main/java/com/smallredtracktor/yourpersonaleducationalapplication/main/Repos/Repositories/CreateTestMemoryRepository.java
@@ -3,23 +3,25 @@ package com.smallredtracktor.yourpersonaleducationalapplication.main.Repos.Repos
 import com.smallredtracktor.yourpersonaleducationalapplication.main.DataObjects.POJOs.OcrResponseModel;
 import com.smallredtracktor.yourpersonaleducationalapplication.main.DataObjects.TestItem;
 import com.smallredtracktor.yourpersonaleducationalapplication.main.LocalDataSources.ICreateTestDbApi;
+import com.smallredtracktor.yourpersonaleducationalapplication.main.LocalDataStorages.ILocalStorage;
 import com.smallredtracktor.yourpersonaleducationalapplication.main.Repos.Interfaces.ICreateTestRepository;
 import com.smallredtracktor.yourpersonaleducationalapplication.main.Utils.PhotoUtils.ParseTextUtil;
 
 import java.util.List;
-
-
 import io.reactivex.Flowable;
-import io.reactivex.Single;
+import io.reactivex.Maybe;
+
 
 public class CreateTestMemoryRepository implements ICreateTestRepository {
 
     private ParseTextUtil util;
     private ICreateTestDbApi testDbApi;
+    private ILocalStorage localStorage;
 
-    public CreateTestMemoryRepository(ICreateTestDbApi testDbApi, ParseTextUtil util) {
+    public CreateTestMemoryRepository(ICreateTestDbApi testDbApi, ParseTextUtil util, ILocalStorage localStorage) {
         this.testDbApi = testDbApi;
         this.util = util;
+        this.localStorage = localStorage;
     }
 
 
@@ -30,16 +32,43 @@ public class CreateTestMemoryRepository implements ICreateTestRepository {
 
     @Override
     public void deleteTestItem(String id) {
-        testDbApi.deleteTestItem(id);
+        getTestItem(id)
+                .toObservable()
+                .doOnNext(testItems -> {
+                    try {
+                        if (testItems.get(0).getId() != null) {
+                            int type = testItems.get(0).getType();
+                            if (type == 1 || type == 2) {
+                                String filepath = testItems.get(0).getValue();
+                                try {
+                                    deleteFile(filepath);
+                                } catch (Exception e) {
+                                    e.printStackTrace();
+                                }
+                            }
+                            testDbApi.deleteTestItem(id);
+                        }
+                    } catch (ArrayIndexOutOfBoundsException e)
+                    {
+                        //Caused by Realm live objects
+                    }
+
+        }).subscribe();
     }
 
     @Override
-    public Single<OcrResponseModel> getParsedTextFromFile(String mPath) {
+    public void deleteFile(String filepath)
+    {
+        localStorage.deleteFile(filepath);
+    }
+
+    @Override
+    public Maybe<OcrResponseModel> getParsedTextFromFile(String mPath) {
         return  util.getResult(mPath);
     }
 
     @Override
-    public void writeTestItem(String id, boolean isQuestion, int currentTicket, int type, String value) {
+    public void writeTestItem(String id, boolean isQuestion, String currentTicket, int type, String value) {
         testDbApi.updateTestItem(id,  isQuestion,  currentTicket, type,  value);
     }
 }
